docs: MainPage -> loginPage
