+++ src/page/HomePage.tsx
@@ -1,0 +1,1 @@
+export default function LoginPage() {}
